backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.5
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0